==> 01-data-manipulation-scripts/02-data-cleaning/07-substrate/07-resample-substrate.Rmd <==
---
title: "Resample Substrate"
author: "Elke Windschitl"
date: "2023-04-18"
output: html_document
---

Load necessary libraries

```{r}
library(tidyverse)
library(raster) 
library(terra)
library(tmap)
library(tmaptools)
library(sf)

```

Set data directory

```{r}
#---- Set up the data directory (wherever your download of our Google Shared Drive lives)
data_dir <- "/capstone/kelpgeomod/google_drive_download"
```

Load data

```{r}
mask <- terra::rast(file.path(data_dir, "02-intermediate-data/03-mask-intermediate/main-mask.tif"))

substrate <- terra::rast(file.path(data_dir, "/02-intermediate-data/08-substrate-intermediate/combined-sbchannel-substrate.tif"))
```

Make binary sandy/non-sandy

```{r}
# Set up a reclassification matrix to identify sandy-bottom only
m <- c(0, 1,
       1, 0,
       2, 0,
       3, 0)
rclmat <- matrix(m, ncol=2, byrow=TRUE)

# Apply the reclassification, name this new raster rc1
rc1 <- classify(substrate, rclmat, include.lowest=TRUE)

#--------------------exploring if the reclass worked-----------------------

# Define the extent for subsetting
subset_extent <- extent(c(-120.5, -120.4, 34.4, 34.5)) 

# Crop og and new to subset for faster processing for investigation
cropped <- crop(rc1,subset_extent) # this is the new
cropped2 <- crop(substrate, subset_extent) # thisis the og

# Make a data frame to compare exact grid cells 
df <- as.data.frame(cropped, xy = TRUE, na.rm = TRUE) # this is the new
df2 <- as.data.frame(cropped2, xy = TRUE, na.rm = TRUE) # this is the og

# Plot to check if expected
plot(rc1)
```

Aggregate and remove non-sandy areas

```{r}
# Check the crs, and reset the crs if different
crs(rc1) == crs(mask)

small <- res(rc1)[1] # grab our smallest resolution
large <- res(mask)[2] # grab our largest resoultion

# Calculate the aggregation factor
factor <- c(large / small, large / small)  # Divide the target resolution (1 km) by the source resolution (3 m)

#Aggregate to the mask size (roughly)
agg_sub <- terra::aggregate(x = rc1,
                            fact = factor, # number of cells in each direction
                            fun = "mean")

# resample to match mask exactly
resampled <- terra::resample(agg_sub, mask, method = "near")

# set up a new classification matrix
# any values <1 had a non-soft cell factor into the mean. We want pure soft-bottom
m <- c(0, 1, NA,
       1, 2, 1)
rclmat2 <- matrix(m, ncol=3, byrow=TRUE)

# Apply the reclassification, name this new raster rc2
rc2 <- classify(resampled, rclmat2, right = FALSE)

# Pull in polygon data to check our soft-substrate raster
polygons <- st_read(file.path(data_dir, "02-intermediate-data/08-substrate-intermediate/usgs-substrate-shape/usgs-substrate.shp")) %>% 
  filter(class == "h" | class == "m" | class == "a")

# Plot our soft substrate raster vs other substrate polygons to look for overlap
tm_shape(rc2) +
  tm_raster() +
  tm_shape(polygons) +
  tm_polygons() +
  tmap_options(check.and.fix = TRUE) 
```

```{r}
#---- Write to data files
# Change this file path to your own to re-write data
sandy_raster <- rc2
writeRaster(sandy_raster, "/capstone/kelpgeomod/google_drive_download/03-analysis-data/05-substrate-analysis/sandy-bottom-1km.tif", overwrite = TRUE)
```
